refactor code out of the UserInfoStore and into a handler
